Added tests

+++ services/rollout-service/pkg/argo/argo.go
@@ -82,16 +82,19 @@ func (a *ArgoAppProcessor) Consume(ctx context.Context, hlth *setup.HealthReport
 	hlth.ReportReady("event-consuming")
 	l := logger.FromContext(ctx).With(zap.String("self-manage", "consuming"))
 	appsKnownToArgo := map[string]map[string]*v1alpha1.Application{}
+	envAppsKnownToArgo := make(map[string]*v1alpha1.Application)
 	for {
 		select {
 		case overview := <-a.trigger:
 			for _, envGroup := range overview.EnvironmentGroups {
 				for _, env := range envGroup.Environments {
-					envAppsKnownToArgo := appsKnownToArgo[env.Name]
-					err := a.DeleteArgoApps(ctx, envAppsKnownToArgo, env.Applications)
-					if err != nil {
-						l.Error("deleting applications")
-						continue
+					if ok := appsKnownToArgo[env.Name]; ok != nil {
+						envAppsKnownToArgo = appsKnownToArgo[env.Name]
+						err := a.DeleteArgoApps(ctx, envAppsKnownToArgo, env.Applications)
+						if err != nil {
+							l.Error("deleting applications")
+							continue
+						}
 					}
 
 					for _, app := range env.Applications {
@@ -128,17 +131,16 @@ func (a ArgoAppProcessor) CreateOrUpdateApp(ctx context.Context, overview *api.G
 	t := team(overview, app.Name)
 	var existingApp *v1alpha1.Application
 	if a.ManageArgoAppsEnabled && len(a.ManageArgoAppsFilter) > 0 && slices.Contains(a.ManageArgoAppsFilter, t) {
-		k := Key{AppName: app.Name, EnvName: env.Name, Application: app, Environment: env}
 
 		for _, argoApp := range appsKnownToArgo {
-			if argoApp.Name == app.Name && argoApp.Annotations["com.freiheit.kuberpult/application"] != "" {
+			if argoApp.Name == fmt.Sprintf("%s-%s", env.Name, app.Name) && argoApp.Annotations["com.freiheit.kuberpult/application"] != "" {
 				existingApp = argoApp
 				break
 			}
 		}
 
 		if existingApp == nil {
-			appToCreate := CreateArgoApplication(overview, app, k.Environment)
+			appToCreate := CreateArgoApplication(overview, app, env)
 			appToCreate.ResourceVersion = ""
 			upsert := false
 			validate := false
@@ -158,7 +160,7 @@ func (a ArgoAppProcessor) CreateOrUpdateApp(ctx context.Context, overview *api.G
 				}
 			}
 		} else {
-			appToUpdate := CreateArgoApplication(overview, app, k.Environment)
+			appToUpdate := CreateArgoApplication(overview, app, env)
 			appUpdateRequest := &application.ApplicationUpdateRequest{
 				XXX_NoUnkeyedLiteral: struct{}{},
 				XXX_unrecognized:     nil,
@@ -167,7 +169,8 @@ func (a ArgoAppProcessor) CreateOrUpdateApp(ctx context.Context, overview *api.G
 				Application:          appToUpdate,
 				Project:              ptr.FromString(appToUpdate.Spec.Project),
 			}
-			if !cmp.Equal(appUpdateRequest.Application.Spec, existingApp.Spec) {
+			//We have to exclude the unexported type isServerInferred. It is managed by Argo.
+			if !cmp.Equal(appUpdateRequest.Application.Spec, existingApp.Spec, cmp.AllowUnexported(v1alpha1.ApplicationSpec{}.Destination)) {
 				_, err := a.ApplicationClient.Update(ctx, appUpdateRequest)
 				if err != nil {
 					logger.FromContext(ctx).Error("updating application: "+appToUpdate.Name+",env "+env.Name, zap.Error(err))
